+++ akshare/__init__.py
@@ -2039,9 +2039,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.6.18 fix: fix stock_zh_a_spot_em interface
 1.6.19 fix: fix stock_comment_detail_zlkp_jgcyd_em interface
 1.6.20 fix: fix stock_board_industry_name_em and code_id_map_em interface
+1.6.21 fix: fix baidu_search_index interface
 """
 
-__version__ = "1.6.20"
+__version__ = "1.6.21"
 __author__ = "AKFamily"
 
 import sys