+++ docs/changelog.md
@@ -29,6 +29,12 @@
 
 ## 更新说明
 
+1.6.21 fix: fix baidu_search_index interface
+
+    1. 修复 baidu_search_index 接口
+    2. 修复 baidu_info_index 接口
+    3. 修复 baidu_media_index 接口
+
 1.6.20 fix: fix stock_board_industry_name_em and code_id_map_em interface
 
     1. 修复 stock_board_industry_name_em 接口，字段容错处理
@@ -765,6 +771,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.6.21 fix: fix baidu_search_index interface
 
 1.6.20 fix: fix stock_board_industry_name_em and code_id_map_em interface